Split up test files for maintainability
