+++ server/api/tests/permissions/test_permission_updates.py
@@ -1,7 +1,7 @@
 from django.urls import reverse
 from rest_framework.test import APITestCase
 from rest_framework import status
-from ..models import MemberPermissions, WorkspaceMember, Workspace, User
+from ...models import MemberPermissions, WorkspaceMember, Workspace, User
 
 
 class UpdatePermissionsTests(APITestCase):
